search_service/docker/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
distro==1.8.0
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
environ==1.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
inflection==0.5.1
Markdown==3.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
